Skill add/delete security check

+++ src/AppBundle/Controller/SkillController.php
@@ -7,6 +7,7 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Component\HttpFoundation\Request;
 use AppBundle\Entity\Skill;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 class SkillController extends Controller
 {
@@ -14,10 +15,11 @@ class SkillController extends Controller
     public function createAction(Request $request) {
 
         $skillTitle = $request->request->get('skill');
-        $userID = (integer)$request->request->get('user-id');
+        //$userID = (integer)$request->request->get('user-id');
 
-        $repository = $this->getDoctrine()->getRepository('AppBundle:UserSeeker');
-        $user = $repository->find($userID);
+        //$repository = $this->getDoctrine()->getRepository('AppBundle:UserSeeker');
+        //$user = $repository->find($userID);
+        $user = $this->getUser();
 
         $skill = new Skill();
         $skill->setTitle($skillTitle);
@@ -44,7 +46,10 @@ class SkillController extends Controller
      */
     public function deleteAction(Request $request, Skill $skill)
     {
-        // TODO: security! check user!
+        // Security! Check user.
+        if ($skill->getUser() !== $this->getUser())
+            throw new AccessDeniedException();
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($skill);
         $em->flush();
